Initial commit.
The initial commit includes an implementation of k-means clustering
which can cluster datasets which are too large to fit into memory.
I tried several different implementations of k-means clustering online
before building this, but the ones that I tried weren't capable of
handling datasets that were large enough to fit on disk but not in
memory.