Switch to using stabilization check.
Also added a detailed comment on a deep insight about
how the stabilization being achieved is secretly a
bitmask that indicates that the selection criteria for
what things need further iterative refinement is the empty
set selection. I think it is a really pretty idea. In
practice it makes the completion check much more complicated
but in terms of the conceptual framework around the algorithm
it means we are doing the same thing everything time. There is
no if statement. There was just nothing left to do.