Fix a subtle markov chain monte carlo sampling bug
When sampling from a markov chain without replacement
with a small dataset but a large chain size we can deplete
the dataset. Since the chain size determines the number of
iterations this leads to an error.

In addition when sampling from a markov chain that is weighted
rather than uniform the sampling method was returning the weights
in a biased way with respect to weights. So future iterations weren't
a sample from the same distribution if we precomputed our entire sample
in the first sampling pass and reused it in the next pass.