Update weight fn (#2)
* Make chain length decisions easier.

AFK and MC2 both require a chain length parameter :m
to be set and the choice of this number has some influence
over the accuracy of the results.

This update helps the user choose a good value for :m through two
different means.  The first is by doing a sanity check on :m during
setup when it is provided to try and catch cases where :m is not a
reasonable value and warning the user if the sanity check is failed.

The second is by providing a default chain length computed on the basis
of the size estimate of the dataset and k.  The value of :m is chosen
on the basis of Bachem's 2016 paper - it is the m that he computes
theoretical guarantees for.

Expert users may be better off doing their own analysis to determine a
stronger :m.

* Update k means sampling weight fn.