Refactors and greatly expands unit tests.
There was a major error in the previous version of the code.
When calculating distances in order to determine the closest
centroid to an assignment the data was loaded into a vector.
The field ordering on the vector was not always the same as
the field ordering of centroids loaded into a vector.  So the
distance calculations were incorrect.

This has been fixed.  The work to fix it also moves the code
closer to allowing the user to specify what columns they want to
cluster on rather than clustering on all columns in the provided
dataset.

More important is how this release strives to ensure that this
error does not happen again.  The code was modified to have a
functional core with an imperative shell.  Unit tests and specs
were added to the functional core.  Additional unit tests were
added testing the imperative shell.

One of the new test cases used to fail because of the error, but
isn't failing anymore.  So at least the assignment issue seems to
be resolved.

I plan to continue to refactor and add more tests in order to
increase the confidence that the code clusters correctly.